+++ index.d.ts
@@ -9,8 +9,16 @@ export interface Credentials {
   password: string
 }
 
-export interface User {
+export interface RegisterCredentials {
   username: string
+  password: string
+  passwordConfirm: string
+}
+
+export interface User {
+  _id: string
+  username: string
+  passwordUpdated: string
 }
 
 export interface ResponseError {