+++ src/constants/errors.ts
@@ -1,4 +1,5 @@
 export default {
   SERVER: 'Server error',
-  USER_NOT_FOUND: 'User not found'
+  USER_NOT_FOUND: 'User not found',
+  USER_EXIST: 'User exist'
 }