+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 0.0.2 (May 20, 2022)
+
+- Login method updated
+- Register method added
+
+
 ### 0.0.1 (May 20, 2022)
 
-Initial core added [#1]
+- Initial core added [#1]